requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
django-mathfilters==1.0.0
djangorestframework==3.14.0
Markdown==3.4.3
Pillow==10.0.0
pytz==2023.3
reportlab==4.0.4
sqlparse==0.4.4
